Create a Cryptocurrency Trading Algorithm
The prominence of Cryptocurrencies as a means of transaction and investment vehicle has seen astronomical growth in recent years - but what is it and how is it different than the traditional currencies we're used to? The currencies you are familar with and likely use on a day-to-day basis, such as the US Dollar and the Euro, are known as fiat currencies. They are issued by the government and there exists a central authority, such as a central bank, to regulate and govern it's use. In stark contrast, a defining feature of a Cryptocurrency is that there is no central regulatory authority. Instead, transactions are held in a ledger or database which is secured cryptographically. Both types of currencies have their advantages and disadvantages.

If, however, you prefer a little more of a hands-off approach, a trading algorithm may be here to save the day. Algorithmic trading is a way to automate making decisions on when to buy or sell an asset based on some pre-programmed instructions and criteria. You design and implement a trading strategy and let the computer do the work for you! If the strategy is effective and well-tested, then you can potentially earn money while you sleep without having to spend hours staring at charts.

Fetch Cryptocurrency market prices and data
Perform basic analysis of market behaviour
Implement a simple algorithmic trading strategy
Analyze the performance of the trading algorithm